Update file structure of payment tests
