+++ saleor/graphql/payment/tests/mutations/test_payment_refund.py
@@ -1,0 +1,87 @@
+import graphene
+
+from saleor.graphql.tests.utils import get_graphql_content
+from saleor.payment import TransactionKind
+from saleor.payment.models import ChargeStatus
+
+REFUND_QUERY = """
+    mutation PaymentRefund($paymentId: ID!, $amount: PositiveDecimal!) {
+        paymentRefund(paymentId: $paymentId, amount: $amount) {
+            payment {
+                id,
+                chargeStatus
+            }
+            errors {
+                field
+                message
+            }
+        }
+    }
+"""
+
+
+def test_payment_refund_success(
+    staff_api_client, permission_manage_orders, payment_txn_captured
+):
+    payment = payment_txn_captured
+    payment.charge_status = ChargeStatus.FULLY_CHARGED
+    payment.captured_amount = payment.total
+    payment.save()
+    payment_id = graphene.Node.to_global_id("Payment", payment.pk)
+
+    variables = {"paymentId": payment_id, "amount": str(payment.total)}
+    response = staff_api_client.post_graphql(
+        REFUND_QUERY, variables, permissions=[permission_manage_orders]
+    )
+    content = get_graphql_content(response)
+    data = content["data"]["paymentRefund"]
+    assert not data["errors"]
+    payment.refresh_from_db()
+    assert payment.charge_status == ChargeStatus.FULLY_REFUNDED
+    assert payment.transactions.count() == 2
+    txn = payment.transactions.last()
+    assert txn.kind == TransactionKind.REFUND
+
+
+def test_payment_refund_with_invalid_argument(
+    staff_api_client, permission_manage_orders, payment_txn_captured
+):
+    payment = payment_txn_captured
+    payment.charge_status = ChargeStatus.FULLY_CHARGED
+    payment.captured_amount = payment.total
+    payment.save()
+    payment_id = graphene.Node.to_global_id("Payment", payment.pk)
+
+    variables = {"paymentId": payment_id, "amount": 0}
+    response = staff_api_client.post_graphql(
+        REFUND_QUERY, variables, permissions=[permission_manage_orders]
+    )
+    content = get_graphql_content(response)
+    data = content["data"]["paymentRefund"]
+    assert len(data["errors"]) == 1
+    assert data["errors"][0]["message"] == "Amount should be a positive number."
+
+
+def test_payment_refund_error(
+    staff_api_client, permission_manage_orders, payment_txn_captured, monkeypatch
+):
+    payment = payment_txn_captured
+    payment.charge_status = ChargeStatus.FULLY_CHARGED
+    payment.captured_amount = payment.total
+    payment.save()
+    payment_id = graphene.Node.to_global_id("Payment", payment.pk)
+    variables = {"paymentId": payment_id, "amount": str(payment.total)}
+    monkeypatch.setattr("saleor.payment.gateways.dummy.dummy_success", lambda: False)
+    response = staff_api_client.post_graphql(
+        REFUND_QUERY, variables, permissions=[permission_manage_orders]
+    )
+    content = get_graphql_content(response)
+    data = content["data"]["paymentRefund"]
+
+    assert data["errors"] == [{"field": None, "message": "Unable to process refund"}]
+    payment.refresh_from_db()
+    assert payment.charge_status == ChargeStatus.FULLY_CHARGED
+    assert payment.transactions.count() == 2
+    txn = payment.transactions.last()
+    assert txn.kind == TransactionKind.REFUND
+    assert not txn.is_success
