+++ saleor/graphql/payment/tests/mutations/test_payment_void.py
@@ -1,0 +1,62 @@
+import graphene
+
+from .....payment import ChargeStatus, TransactionKind
+from ....tests.utils import get_graphql_content
+
+VOID_QUERY = """
+    mutation PaymentVoid($paymentId: ID!) {
+        paymentVoid(paymentId: $paymentId) {
+            payment {
+                id,
+                chargeStatus
+            }
+            errors {
+                field
+                message
+            }
+        }
+    }
+"""
+
+
+def test_payment_void_success(
+    staff_api_client, permission_manage_orders, payment_txn_preauth
+):
+    assert payment_txn_preauth.charge_status == ChargeStatus.NOT_CHARGED
+    payment_id = graphene.Node.to_global_id("Payment", payment_txn_preauth.pk)
+    variables = {"paymentId": payment_id}
+    response = staff_api_client.post_graphql(
+        VOID_QUERY, variables, permissions=[permission_manage_orders]
+    )
+    content = get_graphql_content(response)
+    data = content["data"]["paymentVoid"]
+    assert not data["errors"]
+    payment_txn_preauth.refresh_from_db()
+    assert payment_txn_preauth.is_active is False
+    assert payment_txn_preauth.transactions.count() == 2
+    txn = payment_txn_preauth.transactions.last()
+    assert txn.kind == TransactionKind.VOID
+
+
+def test_payment_void_gateway_error(
+    staff_api_client, permission_manage_orders, payment_txn_preauth, monkeypatch
+):
+    assert payment_txn_preauth.charge_status == ChargeStatus.NOT_CHARGED
+    payment_id = graphene.Node.to_global_id("Payment", payment_txn_preauth.pk)
+    variables = {"paymentId": payment_id}
+    monkeypatch.setattr("saleor.payment.gateways.dummy.dummy_success", lambda: False)
+    response = staff_api_client.post_graphql(
+        VOID_QUERY, variables, permissions=[permission_manage_orders]
+    )
+    content = get_graphql_content(response)
+    data = content["data"]["paymentVoid"]
+    assert data["errors"]
+    assert data["errors"][0]["field"] is None
+    assert data["errors"][0]["message"] == "Unable to void the transaction."
+    payment_txn_preauth.refresh_from_db()
+    assert payment_txn_preauth.charge_status == ChargeStatus.NOT_CHARGED
+    assert payment_txn_preauth.is_active is True
+    assert payment_txn_preauth.transactions.count() == 2
+    txn = payment_txn_preauth.transactions.last()
+    assert txn.kind == TransactionKind.VOID
+    assert not txn.is_success
