+++ saleor/graphql/payment/tests/queries/test_query_payment.py
@@ -1,0 +1,74 @@
+import graphene
+
+from ....tests.utils import get_graphql_content, get_graphql_content_from_response
+
+QUERY_PAYMENT_BY_ID = """
+    query payment($id: ID!) {
+        payment(id: $id) {
+            id
+            checkout {
+                token
+            }
+        }
+    }
+"""
+
+
+def test_query_payment(payment_dummy, user_api_client, permission_manage_orders):
+    query = QUERY_PAYMENT_BY_ID
+    payment = payment_dummy
+    payment_id = graphene.Node.to_global_id("Payment", payment.pk)
+    variables = {"id": payment_id}
+    response = user_api_client.post_graphql(
+        query, variables, permissions=[permission_manage_orders]
+    )
+    content = get_graphql_content(response)
+    payment_data = content["data"]["payment"]
+    received_id = payment_data["id"]
+    assert received_id == payment_id
+    assert not payment_data["checkout"]
+
+
+def test_query_payment_with_checkout(
+    payment_dummy, user_api_client, permission_manage_orders, checkout
+):
+    query = QUERY_PAYMENT_BY_ID
+    payment = payment_dummy
+    payment.order = None
+    payment.checkout = checkout
+    payment.save()
+    payment_id = graphene.Node.to_global_id("Payment", payment.pk)
+    variables = {"id": payment_id}
+    response = user_api_client.post_graphql(
+        query, variables, permissions=[permission_manage_orders]
+    )
+    content = get_graphql_content(response)
+    payment_data = content["data"]["payment"]
+    received_id = payment_data["id"]
+    assert received_id == payment_id
+    assert payment_data["checkout"]["token"] == str(checkout.pk)
+
+
+def test_staff_query_payment_by_invalid_id(
+    staff_api_client, payment_dummy, permission_manage_orders
+):
+    id = "bh/"
+    variables = {"id": id}
+    response = staff_api_client.post_graphql(
+        QUERY_PAYMENT_BY_ID, variables, permissions=[permission_manage_orders]
+    )
+    content = get_graphql_content_from_response(response)
+    assert len(content["errors"]) == 1
+    assert content["errors"][0]["message"] == f"Couldn't resolve id: {id}."
+    assert content["data"]["payment"] is None
+
+
+def test_staff_query_payment_with_invalid_object_type(
+    staff_api_client, payment_dummy, permission_manage_orders
+):
+    variables = {"id": graphene.Node.to_global_id("Order", payment_dummy.pk)}
+    response = staff_api_client.post_graphql(
+        QUERY_PAYMENT_BY_ID, variables, permissions=[permission_manage_orders]
+    )
+    content = get_graphql_content(response)
+    assert content["data"]["payment"] is None
